Successfully save files in pcm

+++ DoneProject/AQRecorder.h
@@ -70,7 +70,7 @@ public:
 	void			StartRecord(CFStringRef inRecordFile);
 	void			StopRecord();		
 	Boolean			IsRunning() const			{ return mIsRunning; }
-	void SaveSamples(UInt32 numSamples, float *buffer);
+	void SaveSamples(UInt32 numSamples, short int *buffer);
 
 	UInt64			startTime;
